Refactor: 메인 페이지 lazy-loading 수정
